+++ src/main/java/com/github/crisdiazalv/billger/infrastructure/rest/dto/account/AccountDTO.java
@@ -11,7 +11,7 @@ public class AccountDTO {
 
     private long id;
     private long user;
-    private long currentBalance;
+    private double currentBalance;
     private String name;
     private String accountNumber;
     private LocalDateTime creationDate;
