+++ src/main/java/com/github/crisdiazalv/billger/infrastructure/rest/dto/account/AccountInfoDTO.java
@@ -9,7 +9,7 @@ public class AccountInfoDTO {
 
     private long id;
     private long user;
-    private long currentBalance;
+    private double currentBalance;
     private String name;
     private String accountNumber;
     private LocalDateTime creationDate;
